requirements.txt
anyio==3.6.2
certifi==2022.12.7
emoji-country-flag==1.3.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
python-dotenv==0.21.1
python-telegram-bot==20.1
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.5.0
